+++ src/main/kotlin/org/caffeine/chaos/Updoot.kt
@@ -3,6 +3,8 @@ package org.caffeine.chaos
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -16,49 +18,20 @@ import java.nio.file.Files
 import java.nio.file.Paths
 import java.nio.file.StandardCopyOption
 import java.util.concurrent.CompletableFuture
+import kotlin.system.exitProcess
 
 @Serializable
-data class GitHubResponse(
+private data class GitHubResponse(
     @SerialName("assets")
     val assets: List<Asset> = listOf(),
-    @SerialName("assets_url")
-    val assetsUrl: String = "",
-    @SerialName("body")
-    val body: String = "",
-    @SerialName("created_at")
-    val createdAt: String = "",
-    @SerialName("draft")
-    val draft: Boolean = false,
-    @SerialName("html_url")
-    val htmlUrl: String = "",
-    @SerialName("id")
-    val id: Int = 0,
-    @SerialName("mentions_count")
-    val mentionsCount: Int = 0,
-    @SerialName("name")
-    val name: String = "",
-    @SerialName("node_id")
-    val nodeId: String = "",
-    @SerialName("prerelease")
-    val prerelease: Boolean = false,
     @SerialName("published_at")
     val publishedAt: String = "",
     @SerialName("tag_name")
     val tagName: String = "",
-    @SerialName("tarball_url")
-    val tarballUrl: String = "",
-    @SerialName("target_commitish")
-    val targetCommitish: String = "",
-    @SerialName("upload_url")
-    val uploadUrl: String = "",
-    @SerialName("url")
-    val url: String = "",
-    @SerialName("zipball_url")
-    val zipballUrl: String = "",
 )
 
 @Serializable
-data class Asset(
+private data class Asset(
     @SerialName("browser_download_url")
     val browserDownloadUrl: String = "",
     @SerialName("content_type")
@@ -83,26 +56,47 @@ data class Asset(
     val url: String = "",
 )
 
-data class Updoot(
+private data class Updoot(
     val clientIsOutOfDate: Boolean,
     val downUrl: String,
     val latestVerString: String,
     val latestVer: Double,
 )
 
-suspend fun updateStatus(client: Client): Updoot {
+suspend fun update(client: Client) = coroutineScope{
+    val updateStatus = updateStatus()
+    if (updateStatus.clientIsOutOfDate) {
+        if (client.config.updater.notify) {
+            log("Client is out of date!", "UPDATER:")
+            log("You are on version $versionString", "UPDATER:")
+            log("Latest version is ${updateStatus.latestVerString}", "UPDATER:")
+            if (!client.config.updater.auto_download) {
+                log("Please update here: ${updateStatus.downUrl}", "UPDATER:")
+            }
+        }
+        if (client.config.updater.auto_download) {
+            downloadUpdate(updateStatus.downUrl).thenAccept {
+                this.launch {
+                    log("Downloaded latest update to ${it}!", "UPDATER:")
+                    if (client.config.updater.exit) {
+                        client.logout()
+                        exitProcess(69)
+                    }
+                }
+            }
+        }
+    }
+}
+
+private suspend fun updateStatus(): Updoot {
 
     //semver fucking rocks
     val git = git()
-    val gnum = git.tagName.replace("[A-z]".toRegex(), "").split(".")
+    val gnum = git.tagName.replace("[^0-9.]".toRegex(), "").split(".")
     val gmajor = gnum[0]
     val gminor = gnum[1]
     val gpatch = gnum[2]
-    val gver = if (gnum[3].toIntOrNull() != null && client.config.updater.prereleases) {
-        "$gmajor.$gminor$gpatch${gnum[3]}".toDouble()
-    } else {
-        "$gmajor.$gminor$gpatch".toDouble()
-    }
+    val gver = "$gmajor.$gminor$gpatch".toDouble()
 
     var downUrl = ""
     for (i in git.assets) {
@@ -113,18 +107,11 @@ suspend fun updateStatus(client: Client): Updoot {
         }
     }
 
-    var clientIsOutOfDate = false
-
-    if (versionDouble != gver) {
-        clientIsOutOfDate = true
-        if (git.prerelease && !client.config.updater.prereleases) {
-            clientIsOutOfDate = false
-        }
-    }
+    val clientIsOutOfDate = versionDouble != gver
     return Updoot(clientIsOutOfDate, downUrl, gnum.joinToString("."), gver)
 }
 
-suspend fun downloadUpdate(url: String): CompletableFuture<String> {
+private suspend fun downloadUpdate(url: String): CompletableFuture<String> {
     val filename = "CHAOS-${versionString}.jar"
     withContext(Dispatchers.IO) {
         val inputStream = URL(url).openStream()
